Change .innerHTML to .innerText to mitigate XSS
An attacker may gain access to the database and inject malicious scripts into the datetime field, leading to malicious code execution in a users browser. While unlikely, this should still be mitigated by modifying the DOM with innerText rather than innerHTML. Currently, the datetime DOM modification will work exactly the same using innerText as innerHTML, so this change should be compatible with the current commit.